refactor: move out command registration
Update commands are left behind as they are dependent on the UpdateExplorer instance being
being available. init and completions generation are left as they are dependent on functions
defined in extension.ts. They will all be moved later

refactor: correct registerCommand usage
