--- src/commands.ts
@@ -1,10 +1,0 @@
-export * from './commands/common';
-
-export * from './commands/alloyGenerate';
-export * from './commands/buildApp';
-export * from './commands/buildModule';
-export * from './commands/packageApp';
-export * from './commands/packageModule';
-export * from './commands/createApp';
-export * from './commands/createModule';
-export * from './commands/clean';
